Refactor: Split amrex module into particle_data and plotting
- Moved the `AMReXParticleData` and `AMReXParticleHeader` classes to `src/flekspy/amrex/particle_data.py`.
- Created a new `AMReXPlottingMixin` class in `src/flekspy/amrex/plotting.py` to house all plotting-related methods.
- Updated `AMReXParticleData` to inherit from `AMReXPlottingMixin`, combining the data and plotting functionalities.
- Adjusted `src/flekspy/amrex/__init__.py` to expose the refactored classes.
- Updated unit tests to reflect the new module structure.
